Added Abouts Page And Remaining Features of HomePage
