--- Benjamin/helpops_3.0/components/ParticleBackground.js
@@ -1,97 +1,0 @@
-import Particles, { initParticlesEngine } from "@tsparticles/react";
-import { useEffect, useMemo, useState } from "react";
-import { loadSlim } from "@tsparticles/slim";
-
-const ParticlesComponent = (props) => {
-  const [init, setInit] = useState(false);
-  useEffect(() => {
-    initParticlesEngine(async (engine) => {
-      await loadSlim(engine);
-    }).then(() => {
-      setInit(true);
-    });
-  }, []);
-
-  const particlesLoaded = (container) => {
-    console.log(container);
-  };
-
-  const options = useMemo(
-    () => ({
-      fpsLimit: 120,
-      interactivity: {
-        events: {
-          onClick: {
-            enable: true,
-            mode: "repulse",
-          },
-          onHover: {
-            enable: true,
-            mode: "repulse",
-          },
-        },
-        modes: {
-          repulse: {
-            distance: 200,
-            duration: 0.4,
-          },
-          grab: {
-            distance: 150,
-          },
-        },
-      },
-      particles: {
-        number: {
-          value: 100,
-          density: {
-            enable: true,
-            value_area: 800,
-          },
-        },
-        color: {
-          value: "#FFFFFF",
-        },
-        shape: {
-          type: "triangle",
-          stroke: {
-            width: 0,
-            color: "#000000",
-          },
-          polygon: {
-            nb_sides: 12,
-          },
-        },
-        links: {
-          color: "#FFFFFF",
-          distance: 180,
-          enable: true,
-          opacity: 0.8,
-          width: 1,
-        },
-        move: {
-          direction: "none",
-          enable: true,
-          outModes: {
-            default: "bounce",
-          },
-          random: true,
-          speed: 5,
-          straight: false,
-        },
-        opacity: {
-          value: 0.7,
-        },
-
-        size: {
-          value: { min: 1, max: 3 },
-        },
-      },
-      detectRetina: true,
-    }),
-    []
-  );
-
-  return <Particles id={props.id} init={particlesLoaded} options={options} />;
-};
-
-export default ParticlesComponent;
