--- Benjamin/helpops_3.0/jsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "paths": {
-      "@*": ["./*"]
-    }
-  }
-}
