--- Benjamin/helpops_3.0/src/app/page.js
@@ -1,13 +1,0 @@
-"use client";
-import Header from "@components/Header";
-import ParticlesComponent from "@components/ParticleBackground";
-import HomePage from "@components/HomePage";
-export default function Home() {
-  return (
-    <div>
-      <Header />
-      <ParticlesComponent />
-      <HomePage/>
-    </div>
-  );
-}
