+++ Benjamin/jsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "paths": {
+      "@*": ["./*"]
+    }
+  }
+}
