+++ Benjamin/src/app/abouts/page.js
@@ -1,0 +1,12 @@
+"use client";
+
+import AboutPage from "@pages/AboutPage";
+
+
+export default function Abouts() {
+  return (
+    <div>
+      <AboutPage />
+    </div>
+  );
+}
