+++ Benjamin/src/app/globals.css
@@ -1,0 +1,10 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+body {
+    margin: 0;
+    font-family: 'Arial', sans-serif;
+    background: linear-gradient(180deg, #FDD86C 8.1%, #FF7D1F 100%);
+    min-height: 100vh;
+}
