+++ Benjamin/src/app/page.js
@@ -1,0 +1,12 @@
+"use client";
+import ParticlesComponent from "@components/ParticleBackground";
+import HomePage from "@pages/HomePage";
+
+export default function Home() {
+  return (
+    <div>
+      <ParticlesComponent />
+      <HomePage />
+    </div>
+  );
+}
