+++ Benjamin/stylesheets/darkmode.css
@@ -1,0 +1,83 @@
+/* darkmode.css */
+
+body.dark-mode {
+  background-color: #121212;
+  color: #e0e0e0;
+  background: linear-gradient(180deg, #333 8.1%, #121212 100%);
+}
+
+.dark-mode .iconsclass > a {
+  width: 35px !important;
+  color: #e0e0e0;
+  padding: 4px !important;
+  border-radius: 50%;
+  font-size: 1.1rem;
+}
+.dark-mode .video{
+  filter: drop-shadow(0 4px 6px #9e9e9e);
+}
+.dark-mode .iconsclass > a:hover {
+  box-shadow: 0 0px 8px rgb(152, 152, 151);
+  color: #b7b7b7;
+}
+
+.dark-mode .font {
+  color: #e0e0e0;
+}
+.dark-mode .nav-links li a::after {
+  background-color: #c0c0c0;
+}
+.dark-mode .active {
+  background-color: #333;
+}
+
+.dark-mode .nav-links {
+  background-color: #333;
+  box-shadow: 0 4px 6px rgba(193, 193, 193, 0.5);
+}
+
+.dark-mode .nav-links li a {
+  color: #e0e0e0;
+}
+
+.dark-mode .sponsor-btn,
+.dark-mode .get-started-btn {
+  background-color: #333;
+  color: #e0e0e0;
+  box-shadow: 0 2px 5px rgba(255, 255, 255, 0.2);
+}
+
+.dark-mode .hamburger div {
+  background-color: #e0e0e0;
+}
+
+.dark-mode p {
+  color: #e0e0e0 !important;
+}
+
+.dark-mode h1 {
+  color: #e0e0e0 !important;
+}
+.dark-mode .text-desc{
+  color: #e0e0e0 !important;
+}
+.dark-mode .about_title {
+  color: #e0e0e0;
+}
+
+.dark-mode .logo img {
+  opacity: 1;
+}
+.drak-mode .rank {
+  background: black;
+}
+.dark-mode #search-box {
+  background-color: #2c2c2c;
+}
+.dark-mode #search-bar {
+  background-color: #2c2c2c;
+  color: #e0e0e0;
+}
+.moon {
+  height: 30px !important;
+}
