+++ Benjamin/stylesheets/footer.css
@@ -1,0 +1,40 @@
+.footer{
+    margin-top: 6rem;
+    gap: 6px;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    align-items: center;
+    flex-direction: column;
+    margin-bottom: 15px;
+}
+
+.iconsclass{
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+}
+.iconsclass >a{
+    width:35px;
+    color: black;
+    padding: 4px;
+    border-radius: 50%;
+    font-size: 1.1rem;
+}
+.iconsclass > a:hover{
+  box-shadow: 0px 0px 8px rgb(255, 255, 255);
+  color: #e0e0e0;
+}
+
+.mm{
+    margin: 0px !important;
+}
+.font{
+    font-size: 0.8rem;
+    font-family: sans-serif;
+    color: black;
+    font-weight:550;
+}
